pipeline/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
snowflake-connector-python==3.14.0
requests==2.32.3
